+++ linopy/constants.py
@@ -194,8 +194,6 @@ class Solution:
     primal: pd.Series = field(default_factory=_pd_series_float)
     dual: pd.Series = field(default_factory=_pd_series_float)
     objective: float = np.nan
-    SAObjUp: pd.Series = field(default_factory=_pd_series_float)
-    SAObjLow: pd.Series = field(default_factory=_pd_series_float)
 
 
 @dataclass